remove  InMemorySignerExt from fiber-lib
